fix(wsClient): Always reconnect when connection fails
